runWIMP: Add Dylan's trigger efficiency SF.
Also moved the ID+ISO SF part so that it is now next to the trigger
efficiency SF. Putting together what belongs together.

The efficiency is binned in tag/probe |eta|. For some reason it is not
symmetric. For now, we take the leading lepton as "tag" and the trailing
one as "probe".